requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
breadability==0.1.20
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
dj-database-url==0.5.0
Django==4.0.4
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-multi-form-view==2.0.1
docopt==0.6.2
gunicorn==20.1.0
idna==3.3
joblib==1.1.0
lxml==4.8.0
networkx==2.8
nltk==3.7
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycountry==22.3.5
PyMuPDF==1.19.6
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
regex==2022.3.15
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
sumy==0.9.0
threadpoolctl==3.1.0
tqdm==4.64.0
urllib3==1.26.9
whitenoise==6.0.0
